(cli): add 'sync' command argument
